Add Tasks endpoint and GetCategories function
A new endpoint for Tasks has been added to the routes. This endpoint uses a new handler, GetTasks, which retrieves distinct task categories from the database. In the database component, a new function, GetCategories, is introduced to select distinct categories from the 'tasks' table. These changes improve the handling of task-related data and make it possible to fetch unique task categories.